+++ src/main/java/com/arcanc/biomorphosis/integration/jei/StomachRecipeCategory.java
@@ -165,7 +165,7 @@ public class StomachRecipeCategory implements IRecipeCategory<RecipeHolder<Stoma
         this.arrow.draw(guiGraphics);
         guiGraphics.pose().popPose();
 
-        guiGraphics.blit(Database.GUI.Textures.JEI.TIME, 5, 60, 0, 0, 8, 8, 16, 16,16, 16);
+        guiGraphics.blit(Database.GUI.Textures.JEI.TIME, 5, 60, 8, 8, 0, 0, 16, 16,16, 16);
         guiGraphics.drawString(RenderHelper.mc().font, Component.literal(Integer.toString(recipe.getResources().time())), 15, 60, 0, false);
     }