+++ changelog.md
@@ -10,3 +10,4 @@
 * added columns to wastes biome
 * added custom font into book (as part of big book rework)
 * fixed jei compatibility. Now recipes registered correctly
+* fixed in book recipe rendering